Ajustando erros na navbar dentro do carrinho

+++ loja/produtos/models.py
@@ -6,7 +6,7 @@ class Addproduto(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(100), nullable=False,unique=True)
     price = db.Column(db.Float, nullable=False)
-    discount = db.Column(db.Float, default=0)
+    discount = db.Column(db.Integer, default=0)
     stock = db.Column(db.Integer, nullable=False)
     description = db.Column(db.String(255), nullable=False)
     color = db.Column(db.String(100), nullable=False)
